refactor(bindings): move tasks to `tasks` folder on nodejs binding
